fixed task embedding type, added curriculum-related vars

+++ nmmo/core/env.py
@@ -48,7 +48,7 @@ class Env(ParallelEnv):
     # Default task: rewards 1 each turn agent is alive
     self.tasks = task_api.nmmo_default_task(self.possible_agents)
     self.agent_task_map = None
-    self._dummy_task_embedding = np.zeros(self.config.TASK_EMBED_DIM, dtype=np.float32)
+    self._dummy_task_embedding = np.zeros(self.config.TASK_EMBED_DIM, dtype=np.float16)
 
     # curriculum file path, if provided, should exist
     self.curriculum_file_path = config.CURRICULUM_FILE_PATH
@@ -73,7 +73,9 @@ class Env(ParallelEnv):
       "AgentId": gym.spaces.Discrete(self.config.PLAYER_N+1),
       "Tile": box(self.config.MAP_N_OBS, Tile.State.num_attributes),
       "Entity": box(self.config.PLAYER_N_OBS, Entity.State.num_attributes),
-      "Task": gym.spaces.Box(low=-2**15, high=2**15-1, shape=(self.config.TASK_EMBED_DIM,)),
+      "Task": gym.spaces.Box(low=-2**15, high=2**15-1,
+                             shape=(self.config.TASK_EMBED_DIM,),
+                             dtype=np.float16),
     }
 
     if self.config.ITEM_SYSTEM_ENABLED:
